+++ pyannote/audio/tasks/segmentation/segmentation.py
@@ -456,7 +456,7 @@ class Segmentation(SegmentationTaskMixin, Task):
 
         # visualize first 9 validation samples of first batch in Tensorboard
         X = X.cpu().numpy()
-        y = y.cpu().numpy()
+        y = y.float().cpu().numpy()
         y_pred = y_pred.cpu().numpy()
         permutated_y_pred = permutated_y_pred.cpu().numpy()
